+++ package.json
@@ -15,19 +15,20 @@
 		"test:synpress:open": "env-cmd -f .env.local synpress open",
 		"test:e2e": "start-server-and-test 'yarn build && yarn start' http-get://localhost:3000 'synpress run'",
 		"test:e2e:ci": "synpress run --group smoke-tests --record --parallel",
-		"upgrade-synthetix": "yarn add @synthetixio/assets@latest @synthetixio/contracts-interface@latest @synthetixio/optimism-networks@latest @synthetixio/providers@latest @synthetixio/queries@latest @synthetixio/transaction-notifier@latest @synthetixio/wei@latest"
+		"upgrade-synthetix": "yarn add @synthetixio/assets@latest @synthetixio/contracts-interface@latest @synthetixio/optimism-networks@latest @synthetixio/providers@latest @synthetixio/queries@latest @synthetixio/transaction-notifier@latest @synthetixio/wei@latest",
+		"upgrade-synthetix:dev": "yarn add @synthetixio/assets@latest @synthetixio/contracts-interface@dev @synthetixio/optimism-networks@dev @synthetixio/providers@dev @synthetixio/queries@dev @synthetixio/transaction-notifier@dev @synthetixio/wei@dev"
 	},
 	"dependencies": {
 		"@artsy/fresnel": "1.8.0",
 		"@reach/dialog": "0.15.3",
 		"@snapshot-labs/snapshot.js": "0.3.22",
 		"@synthetixio/assets": "2.0.12",
-		"@synthetixio/contracts-interface": "2.66.0-1",
-		"@synthetixio/optimism-networks": "2.66.0-1",
-		"@synthetixio/providers": "2.66.0-1",
-		"@synthetixio/queries": "2.66.0-1",
-		"@synthetixio/transaction-notifier": "2.66.0-1",
-		"@synthetixio/wei": "2.66.0-1",
+		"@synthetixio/contracts-interface": "0.0.0-f525272",
+		"@synthetixio/optimism-networks": "0.0.0-f525272",
+		"@synthetixio/providers": "0.0.0-f525272",
+		"@synthetixio/queries": "0.0.0-f525272",
+		"@synthetixio/transaction-notifier": "0.0.0-f525272",
+		"@synthetixio/wei": "0.0.0-f525272",
 		"@tippyjs/react": "4.1.0",
 		"@uniswap/router-sdk": "1.0.5",
 		"@uniswap/sdk-core": "^3.0.1",
